+++ src/scalib/__init__.py
@@ -1,2 +1,2 @@
-__all__ = ["metrics", "attacks", "modeling", "postprocessing"]
+__all__ = ["metrics", "attacks", "modeling", "postprocessing", "config"]
 from .version import version as __version__