+++ src/scalib/config/__init__.py
@@ -1,0 +1,1 @@
+from .threading import thread_context, ThreadPool, default_threadpool