+++ src/scalib_ext/scalib-py/src/ttest.rs
@@ -1,5 +1,6 @@
 //! Python wrapper for SCALib's Ttest
 
+use crate::thread_pool::ThreadPool;
 use numpy::{PyArray1, PyArray2, PyReadonlyArray1, PyReadonlyArray2, ToPyArray};
 use pyo3::prelude::*;
 
@@ -22,17 +23,27 @@ impl Ttest {
     /// Update the Ttest state with n fresh traces
     /// traces: the leakage traces with shape (n,ns)
     /// y: realization of random variables with shape (n,)
-    fn update(&mut self, py: Python, traces: PyReadonlyArray2<i16>, y: PyReadonlyArray1<u16>) {
+    fn update(
+        &mut self,
+        py: Python,
+        traces: PyReadonlyArray2<i16>,
+        y: PyReadonlyArray1<u16>,
+        thread_pool: &ThreadPool,
+    ) {
         let traces = traces.as_array();
         let y = y.as_array();
-        py.allow_threads(|| self.inner.update(traces, y));
+        crate::on_worker(py, thread_pool, || self.inner.update(traces, y));
     }
 
     /// Generate the actual Ttest metric based on the current state.
     /// return array axes (d,ns)
-    fn get_ttest<'py>(&mut self, py: Python<'py>) -> PyResult<&'py PyArray2<f64>> {
-        let ttest = py.allow_threads(|| self.inner.get_ttest());
-        Ok(&(ttest.to_pyarray(py)))
+    fn get_ttest<'py>(
+        &mut self,
+        py: Python<'py>,
+        thread_pool: &ThreadPool,
+    ) -> PyResult<&'py PyArray2<f64>> {
+        let ttest = crate::on_worker(py, thread_pool, || self.inner.get_ttest());
+        Ok(ttest.to_pyarray(py))
     }
 }
 
@@ -56,16 +67,26 @@ impl MTtest {
     /// Update the Ttest state with n fresh traces
     /// traces: the leakage traces with shape (n,ns)
     /// y: realization of random variables with shape (n,)
-    fn update(&mut self, py: Python, traces: PyReadonlyArray2<i16>, y: PyReadonlyArray1<u16>) {
+    fn update(
+        &mut self,
+        py: Python,
+        traces: PyReadonlyArray2<i16>,
+        y: PyReadonlyArray1<u16>,
+        thread_pool: &ThreadPool,
+    ) {
         let traces = traces.as_array();
         let y = y.as_array();
-        py.allow_threads(|| self.inner.update(traces, y));
+        crate::on_worker(py, thread_pool, || self.inner.update(traces, y));
     }
 
     /// Generate the actual Ttest metric based on the current state.
     /// return array axes (d,ns)
-    fn get_ttest<'py>(&mut self, py: Python<'py>) -> PyResult<&'py PyArray1<f64>> {
-        let ttest = py.allow_threads(|| self.inner.get_ttest());
-        Ok(&(ttest.to_pyarray(py)))
+    fn get_ttest<'py>(
+        &mut self,
+        py: Python<'py>,
+        thread_pool: &ThreadPool,
+    ) -> PyResult<&'py PyArray1<f64>> {
+        let ttest = crate::on_worker(py, thread_pool, || self.inner.get_ttest());
+        Ok(ttest.to_pyarray(py))
     }
 }